pyproject.toml
[project]
name = "gepadantic"
version = "0.1.0"
description = "Pydantic-AI GEPA adapter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai-slim[openai, retries, cli]>=1.5.0",
    "gepa>=0.0.12",
    "cloudpickle>=3.0.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "seaborn>=0.13.2",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "mcp-run-python>=0.0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.1",
    "time-machine>=2.19.0",
    "inline-snapshot>=0.29.0",
    "pytest-asyncio>=1.2.0",
]
lint = [
    "ruff>=0.14.6",
]

[tool.inline-snapshot]
# Inline-snapshot expects a format string; provide ruff invocation with filename placeholder.
format-command = "ruff format --stdin-filename {filename} -"